Move common package to op/common
