+++ op/common/image_pull.go
