+++ op/etcd_boot.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/cybozu-go/cke"
-	"github.com/cybozu-go/cke/common"
+	"github.com/cybozu-go/cke/op/common"
 )
 
 type etcdBootOp struct {
