+++ op/kubelet_restart.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/cybozu-go/cke"
-	"github.com/cybozu-go/cke/common"
+	"github.com/cybozu-go/cke/op/common"
 	yaml "gopkg.in/yaml.v2"
 )
