+++ op/proxy_boot.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/cybozu-go/cke"
-	"github.com/cybozu-go/cke/common"
+	"github.com/cybozu-go/cke/op/common"
 	"k8s.io/client-go/tools/clientcmd"
 )
