+++ op/proxy_restart.go
@@ -2,7 +2,7 @@ package op
 
 import (
 	"github.com/cybozu-go/cke"
-	"github.com/cybozu-go/cke/common"
+	"github.com/cybozu-go/cke/op/common"
 )
 
 type kubeProxyRestartOp struct {
